+++ R/utils-control.R
@@ -211,9 +211,10 @@ accepted_controls_by_api <- function(api_type = c("SOAP", "REST", "Bulk 1.0", "B
 #' @note This function is meant to be used internally. Only use when debugging.
 #' @keywords internal
 #' @export
-accepted_controls_by_operation <- function(operation = c("delete", "hardDelete", "insert", "update", "upsert", 
-                                                         "query", "queryall", "retrieve", "resetPassword", 
-                                                         "describeSObjects")){
+accepted_controls_by_operation <- function(operation = c("delete", "undelete", "hardDelete", 
+                                                         "insert", "update", "upsert", 
+                                                         "query", "queryall", "retrieve", 
+                                                         "resetPassword", "describeSObjects")){
   record_creation_controls <- c("AllOrNoneHeader", "AllowFieldTruncationHeader", 
                                 "AssignmentRuleHeader", "DisableFeedTrackingHeader", 
                                 "DuplicateRuleHeader", "EmailHeader", "MruHeader")
@@ -223,7 +224,9 @@ accepted_controls_by_operation <- function(operation = c("delete", "hardDelete",
   switch(this_operation, 
          "delete" = c("AllOrNoneHeader", "DisableFeedTrackingHeader", "EmailHeader", 
                       "UserTerritoryDeleteHeader", bulk_controls),
+         "undelete" = c("AllOrNoneHeader", "AllowFieldTruncationHeader"), 
          "hardDelete" = bulk_controls,
+         "merge" = "AllowFieldTruncationHeader",
          "insert" = c(record_creation_controls, bulk_controls),
          "update" = c(record_creation_controls, bulk_controls, "OwnerChangeOptions"),
          "upsert" = c(record_creation_controls, bulk_controls, "OwnerChangeOptions"),