+++ R/utils.R
@@ -7,6 +7,23 @@ salesforcer_state <- function(){
   .state
 }
 
+#' Return NA if NULL
+#' 
+#' @note This function is meant to be used internally. Only use when debugging.
+#' @keywords internal
+#' @export
+merge_null_to_na <- function(x){
+  if(is.null(x)){ 
+    NA
+  } else if(length(x) == 0){
+    NA
+  } else if(is.list(x) & length(x) == 1 & is.null(x[[1]])){
+    NA
+  } else {
+    x
+  }
+}
+
 #' Write a CSV file in format acceptable to Salesforce APIs
 #' 
 #' @importFrom readr write_csv
@@ -59,8 +76,17 @@ sf_input_data_validation <- function(input_data, operation=''){
   
   # TODO: Add Automatic date validation
   # https://developer.salesforce.com/docs/atlas.en-us.api_bulk_v2.meta/api_bulk_v2/datafiles_date_format.htm
-  
-  if(!is.data.frame(input_data)){
+
+  if(operation == "merge"){
+    # dont try to cast the input for these operations, simply check existence of params
+    list_based <- TRUE
+    stopifnot(names(input_data) == c("victim_ids", "master_fields"))
+  } else if(operation %in% c("getDeleted", "getUpdated")){
+    # dont try to cast the input for these operations, simply check existence of params
+    list_based <- TRUE
+    stopifnot(names(input_data) == c("start", "end"))
+  } else if(!is.data.frame(input_data)){
+    list_based <- FALSE
     if(is.null(names(input_data))){
       if(!is.list(input_data)){
         input_data <- as.data.frame(list(input_data), stringsAsFactors = FALSE)    
@@ -70,22 +96,27 @@ sf_input_data_validation <- function(input_data, operation=''){
     } else {
       input_data <- as.data.frame(as.list(input_data), stringsAsFactors = FALSE)  
     }
-  }
-
-  if(operation %in% c("describeSObjects") & ncol(input_data) == 1){
-    names(input_data) <- "sObjectType"
+  } else {
+    # if already a data.frame, then do nothing
   }
   
-  if(operation %in% c("delete", "retrieve", "findDuplicatesByIds") & ncol(input_data) == 1){
-    names(input_data) <- "Id"
-  }
-  
-  if(operation %in% c("delete", "update", "findDuplicatesByIds")){
-    if(any(grepl("^ID$|^IDS$", names(input_data), ignore.case=TRUE))){
-      idx <- grep("^ID$|^IDS$", names(input_data), ignore.case=TRUE)
-      names(input_data)[idx] <- "Id"
+  # list-based inputs, don't modify, just check existence
+  if(!list_based){
+    if(operation %in% c("describeSObjects") & ncol(input_data) == 1){
+      names(input_data) <- "sObjectType"
     }
-    stopifnot("Id" %in% names(input_data))
+    if(operation %in% c("delete", "undelete", "emptyRecycleBin", 
+                        "retrieve", "findDuplicatesByIds") & ncol(input_data) == 1){
+      names(input_data) <- "Id"
+    }
+    if(operation %in% c("delete", "undelete", "emptyRecycleBin", 
+                        "update", "findDuplicatesByIds")){
+      if(any(grepl("^ID$|^IDS$", names(input_data), ignore.case=TRUE))){
+        idx <- grep("^ID$|^IDS$", names(input_data), ignore.case=TRUE)
+        names(input_data)[idx] <- "Id"
+      }
+      stopifnot("Id" %in% names(input_data))
+    }
   }
   
   return(input_data)
@@ -173,7 +204,7 @@ format_headers_for_verbose <- function(request_headers){
 #' @export
 make_verbose_httr_message <- function(method, url, headers = NULL, body = NULL){
   message(sprintf("%s %s", method, url))
-  if(!is.null(headers)) message(format_headers_for_verbose(headers))
-  if(!is.null(body)) message(sprintf("Body:\n%s", body))
+  if(!is.null(headers)) message(sprintf("\nHeaders\n%s", format_headers_for_verbose(headers)))
+  if(!is.null(body)) message(sprintf("\nBody\n%s", body))
   return(invisible())
 }